classes
The game is seperated into different classes. There is a class for
storing and loading the lexicon. The evil and good gameplay class extend
the general gameplay class.
A game end and highscore is still not implemented. Layouts also are
needed.  And a large wordlist needs to be loaded in